Moved model files to correct location inside project
