+++ src/app/modules/organization/_models/vaccination.model.ts
