+++ src/app/modules/registration-form/general-information/general-information.component.ts
@@ -1,6 +1,6 @@
 import { Component, OnInit } from '@angular/core';
 import { GeneralInformationService } from '../../organization/_services/general-information/general-information.service';
-import { VaccinationModel } from '../_models/vaccination.model';
+import { VaccinationModel } from '../../organization/_models/vaccination.model';
 import { Observable } from 'rxjs';
 
 @Component({
